Replace futures mpsc and oneshot by tokio-sync
The tokio team released the tokio-sync
crate which offers a mpsc and oneshot channel
analogous to the ones in the futures crate.

In some benchs, the new mpsc shows up to 7x improvement
over the futures version. The oneshot implementation
only provides a slight performance improvement, but when
updated to the std version of Future a much better
performance can be expected.

toshi-search#115